requirements.txt
lxml
requests
python-dotenv
easyocr
imageio[pyav]
opencv-python
torch
torchvision
torchaudio
matplotlib
azure-ai-vision-imageanalysis
python-doctr[tf]
mplcursors
pytesseract
pyzbar
sqlalchemy
psycopg2
pandas
seaborn
missingno
fastapi[standard]
passlib
ecdsa-0.19.1 
pyasn1-0.4.8 
pyjwt-2.10.1 
python-jose-3.4.0 
rsa-4.9